Force Field + Playback with present actors
Added a force field that becomes more visible as the target get's closer
to the edge of the field.   A new inverted mesh needs to be added so the
render is visible *inside* the force field.

Playback needs to be improved so that when you stop it playing the
playback it restores control to actors behavior that was disabled.
Things needs to be moved around and loading recordings should be made
static.